Made the SoapClient class configurable.
There are use cases where one have to use an extended SoapClient
class. For example if it is necessary to add support for the NTLM
authentication protocol in order to talk to the SOAP api. [1]

This commit makes the default \SoapClient class configurable through
the Config class. Needless to say, it defaults to \SoapClient to avoid
bc breaks in the first place.

The chosen $clientClass is used for connecting to the api in the first place,
but also for the generated class files.

[1] http://blogs.msdn.com/b/freddyk/archive/2010/01/19/connecting-to-nav-web-services-from-php.aspx