Warn if SOAP_SINGLE_ELEMENT_ARRAYS is not enabled.
To do this PHPUnit has be to be able to mock certain methods in the
Generator class. This is only possible by changing their visibility to
protected. This should also help others who wish to extend the
generator.

PHPUnit also requires that all classes in method signatures are
available so psr/log has been readded as a developer dependency.