Fixed schema referencing when generating code from multiple WSDL files referencing previously loaded XSDs.
When code is generated from multiple WSDL files which share common XSD files,
only the first WSDL would contain the right references to the used XSD files.
This is due to the fact that no references were set in the constructor of the
SchemaDocument class when the referenced XSD file was previously loaded.

This patch creates a link between the referenced XSD url and the initialized
SchemaDocument and will set the used references correctly when the XSD was
already loaded. Also makes sure that schemas referring to each other won't
result in endless loops in SchemaDocument::findTypeElement.